feat: update lydia backend
translate unsupported operators like weak until and strong release
with Lydia grammar
